external/gpl3/gcc:  update generation of the stack protection code                     in function prologues
Refactor stack_protect_prologue() a bit.  When a target provides the
'stack_protect_combined_set' or 'stack_protect_set' instruction
patterns, expand the stack protection guard expression just before
generating and emitting the corresponding instructions.

Otherwise, use expand_assignment() to move a value into the stack
protection slot on the stack.  This will generate code to handle the
case where the stack is aligned to a finer boundary than the size of
the stack protection slot (such as may happen on m68k).